Thank you panel + HEALER
I made a healer the fonction is inside the conversation manager.
I made some debug that I have remove but the cash get remove and the health goes back to max
the problem is that the health bar ui does not increase back to the maximum

I also have an issue with typewriting for the Thank you panel and if you don't have enough money I want to increase the question index to the next one so it said " you don't have enough" I am working on it unless anyone have a solution that won't take a refactoring and 1week.

Check Realisation 2D for a solution for the ending it's quite simple but for the time we have might be a good thing I can do a animation likethe one use in death screen but for gameover with some TEXT that can be modified depending of the outcome of the run.

We need to figure the player health damage starting gold etc

So I can start to do real testing for the presentation